Allow API_URL to be used in the browser
This was only working when API_URL came from a .env file because by default, any OS-level environment variables are required to start with `GATSBY_`.

Instead of adding this whitelist, we could just prefix all environment variables to be used in the browser with `GATSBY_`, but having this whitelist option may come in handy.